requirements.txt
setuptools
pandas
numpy
pymysql
python-dotenv
seaborn
matplotlib
# -e .